requirements.txt
Flask==2.0.1
Flask-Cors==5.0.0
boto3==1.35.23
pandas==2.2.0
numpy==1.25.0
pillow==10.2.0
requests==2.32.3
rasterio==1.3.11
segmentation-models-pytorch==0.3.4
shapely==2.0.6
torch==2.5.0
